pyproject.toml
[tool.poetry]
name = "retail-media-ai"
version = "0.1.0"
description = ""
authors = ["Peter Milne <helipilot50@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<=3.13"

crewai = { extras = ["tools"], version = "^0.51.1" }
python-dotenv = "^1.0.0"
langchain-openai = "^0.1.7"
unstructured = '==0.10.25'
cachetools = "^5.4.0"
decouple = "^0.0.7"
langchain-community = "^0.2.12"

[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"
pytest-mock = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
